move sf and bq to their 'extras' folder
